+++ packages/web3-utils/src/index.ts
@@ -27,3 +27,4 @@ export * from './json_rpc';
 export * as jsonRpc from './json_rpc';
 export * from './web3_deferred_promise';
 export * from './chunk_response_parser';
+export * from './uuid';
